requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crontab==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-rq==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
freezegun==1.5.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.66.3
orjson==3.10.15
packaging==24.2
pillow
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
playwright==1.50.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pycparser==2.22
pyee==12.1.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rq==2.1.0
rq-scheduler==0.14.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
whoosh==2.7.4
yarl==1.18.3
